write tests for ProgressBar, BackButton, and StatsTrack
