+++ test/statsTrack.test.jsx
@@ -1,30 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import renderer from 'react-test-renderer';
 import StatsTrack from '../client/components/statsTrack';
-import ProgressBar from '../client/components/ProgressBar';
-import BackButton from '../client/components/BackButton';
-
-describe('Progress bar', () => {
-  test('Value is rendered correctly', () => {
-    const tree = renderer.create(<ProgressBar fill={32} goal={64} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
-
-describe('Time remaining on campaign', () => {
-  test('Remaining time is read out in days if 1 or more days are left', () => {
-    const statTracker = shallow(<StatsTrack />);
-    statTracker.setState({ deadline: '2019-11-08' });
-    expect(statTracker.find('.units').text()).toEqual('days to go');
-  });
-
-  test('Remaining time is read out in hours if less than 1 day is left', () => {
-    const statTracker = shallow(<StatsTrack />);
-    statTracker.setState({ deadline: '2018-11-01' });
-    expect(statTracker.find('.units').text()).toEqual('hours to go');
-  });
-});
 
 describe('<StatsTrack />', () => {
   it('renders a ProgressBar', () => {
@@ -45,5 +21,26 @@ describe('<StatsTrack />', () => {
   it('renders a pledged amount readout', () => {
     const track = shallow(<StatsTrack />);
     expect(track.find('.pledgedAmount')).toHaveLength(1);
+  });
+
+  it('renders a backers readout', () => {
+    const track = shallow(<StatsTrack />);
+    expect(track.find('.backerCount')).toHaveLength(1);
+  });
+
+  it('displays updated stats when a new pledge is made', () => {
+    const track = shallow(<StatsTrack />);
+    track.setState({
+      backers: 18,
+      pledged: 15876,
+    });
+    const oldBackers = track.find('div.backerCount').text();
+    const oldPledged = track.find('div.pledgedAmount').text();
+    track.setState({
+      backers: 32,
+      pledged: 18745,
+    });
+    expect(track.find('div.backerCount').text()).not.toEqual(oldBackers);
+    expect(track.find('div.pledgedAmount').text()).not.toEqual(oldPledged);
   });
 });
